# specs/001-fullstack-todo-app/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo API
  description: RESTful API for Phase II Todo Application with JWT authentication
  version: 1.0.0
  contact:
    name: Phase II Todo App

servers:
  - url: http://localhost:8000
    description: Development server

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          description: Task title (required)
          example: "Buy groceries"
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Task description (optional)
          example: "Milk, bread, eggs"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          description: Updated task title
          example: "Buy groceries and supplies"
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Updated task description
          example: "Milk, bread, eggs, butter"
        completed:
          type: boolean
          description: Task completion status
          example: true

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 1
        user_id:
          type: string
          description: Owner's user ID
          example: "user_abc123"
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          description: Task description
          example: "Milk, bread, eggs"
        completed:
          type: boolean
          description: Completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-12-30T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-12-30T10:00:00Z"

    TaskListResponse:
      type: array
      items:
        $ref: '#/components/schemas/TaskResponse'

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Task not found"

    DeleteResponse:
      type: object
      properties:
        deleted:
          type: boolean
          example: true

  parameters:
    UserIdParam:
      name: user_id
      in: path
      required: true
      description: User ID (must match authenticated user)
      schema:
        type: string
      example: "user_abc123"

    TaskIdParam:
      name: task_id
      in: path
      required: true
      description: Task ID
      schema:
        type: integer
      example: 1

  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Could not validate credentials"

    Forbidden:
      description: User ID mismatch - cannot access other users' data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Cannot access other users' tasks"

    NotFound:
      description: Task not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Task not found"

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string

paths:
  /api/{user_id}/tasks:
    get:
      summary: List all tasks
      description: Returns all tasks for the authenticated user
      operationId: listTasks
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      operationId: createTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/{user_id}/tasks/{task_id}:
    get:
      summary: Get a single task
      description: Returns a specific task by ID
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update a task
      description: Updates an existing task
      operationId: updateTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/TaskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a task
      description: Permanently deletes a task
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion
      description: Toggles the completion status of a task
      operationId: toggleTaskComplete
      tags:
        - Tasks
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Task completion toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /:
    get:
      summary: Health check
      description: Returns API health status
      operationId: healthCheck
      tags:
        - Health
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  version:
                    type: string
                    example: "1.0.0"

tags:
  - name: Tasks
    description: Task CRUD operations
  - name: Health
    description: API health check